MENG-48: Data processing and network training
Wrote Python scripts to process the Stanford Sentiment140 and Stanford Sentiment Treebank datasets into training and validation sets that are easy to be read and interpreted for training. Wrote a script enabling random accuracy to be calculated. Wrote a script enabling two processed datasets to be combined. Developed a collection of different Neural Networks using Keras and Tensorflow for sentiment analysis. Implemented method to load pre-trained Google Word2Vec embeddings into a Keras model. Generated word to index and index to word mappings to use for the word embeddings.