MENG-48: Generated model, wrote scripts to process datasets
Wrote Python script to expand upon the existing Stanford Sentiment Treebank dataset, growing the training set from ~240,000 data points to ~1.5 million data points. This was done by substituting synonyms within a sentence using WordNet. A model was then chosen, trained and exported for a binary classifier into positive and negative, which achieved an accuracy of 88% on the training set, 81.1% on the validation set, and 81.4% on the test set. The model weights are not able to be included due to the size of the file (3.5 GB), but to achieve this performance the network was trained for 50 epochs which took approximately 8 hours.